chore: skip open() under coding agents
Move isCodingAgent() and guessClientName() from cli-client/utils.ts to
packages/utils/env.ts so reporters and the trace viewer can reach them.
avoid spawning browsers under Claude Code / Copilot CLI.
