+++ packages/playwright-core/src/tools/cli-daemon/program.ts
@@ -20,14 +20,13 @@ import fs from 'fs';
 import os from 'os';
 import path from 'path';
 
-import { getAsBooleanFromENV } from '@utils/env';
+import { getAsBooleanFromENV, guessClientName } from '@utils/env';
 import { libPath } from '../../package';
 import { startCliDaemonServer } from './daemon';
 import { setupExitWatchdog } from '../mcp/watchdog';
 import { createBrowserWithInfo } from '../mcp/browserFactory';
 import * as configUtils from '../mcp/config';
 import { createClientInfo } from '../cli-client/registry';
-import { guessClientName } from '../cli-client/utils';
 import { registry as browserRegistry } from '../../server/registry/index';
 import type { Command } from 'commander';
