+++ packages/utils/DEPS.list
@@ -16,3 +16,6 @@ node_modules/yazl
 [comparators.ts]
 ./image_tools
 ./third_party/pixelmatch
+
+[env.ts]
+"strict"
